Verified with hardware (Changes made)
The previous commit had problems within the Fluid and Drive libraries. The attachment of the Servos in Fluid had to happen in the setup function instead of the Constructor (something, something environment initialisation??). And I forgot to put an abs() around the speed in the the PWM for the Drive H-Bridge, causing it to only work in one direction.